Modified Projectile.cpp and Projectile_BP

+++ Source/BattleTank/Private/TankAimingComponent.cpp
@@ -41,7 +41,7 @@ EFiringState UTankAimingComponent::GetFiringState()
 	return FiringState;
 }
 
-uint8 UTankAimingComponent::GetRoundsLeft()
+int32 UTankAimingComponent::GetRoundsLeft()
 {
 	return Ammo;
 }
